update: separate lattice tests:
